+++ app/Http/Controllers/Admin/Invoice/InvoiceItemController.php
@@ -17,6 +17,7 @@ class InvoiceItemController extends Controller
   {
     $data['invoice'] = $invoice;
     $data['tax_rates'] = Tax::get();
+    $data['is_editable'] = $invoice->isEditable();
 
     if (request()->mode == 'edit') {
       $data['invoice']->load('items.invoiceable');
@@ -46,6 +47,10 @@ class InvoiceItemController extends Controller
 
   public function store(InvoiceItemsStoreReqeust $request, Invoice $invoice)
   {
+    if(!$invoice->isEditable()){
+      return $this->sendError('Invoice is not editable');
+    }
+
     $phases = filterInputIds($request->phases);
 
     $pivot_amounts = ContractPhase::whereIn('id', $phases)
@@ -79,13 +84,26 @@ class InvoiceItemController extends Controller
     return $this->sendRes('Item Added Successfully', ['event' => 'functionCall', 'function' => 'reloadPhasesList', 'close' => 'globalModal']);
   }
 
-  public function destroy(Invoice $invoice, $invoiceItem)
+  public function destroy(Invoice $invoice, $invoiceItem, Request $request)
   {
-    $item = $invoice->items()->find($invoiceItem);
-    if ($item->invoiceable_type == CustomInvoiceItem::class)
-      $item->invoiceable->delete();
+    if(!$invoice->isEditable()){
+      return $this->sendError('Invoice is not editable');
+    }
 
-    $invoice->items()->where('id', $invoiceItem)->delete();
+    $request->validate([
+      'ids' => 'required|array',
+      'ids.*' => 'required|exists:invoice_items,id'
+    ]);
+
+    $items = $invoice->items()->whereIn('invoice_items.id', $request->ids)->get();
+
+    $items->each(function($item){
+      if ($item->invoiceable_type == CustomInvoiceItem::class)
+        $item->invoiceable->delete();
+
+      $item->taxes()->detach();
+      $item->delete();
+    });
 
     $invoice->updateTaxAmount();
